feat(arcgis-rest-portal): split set user properties
